refactor: move session options config to dedicated file
- Move SessionOptionsFromConfig() from utils/tools.go to options.go
- Improves separation of concerns:
  - utils/tools.go: generic utilities (URL, headers, key management)
  - options.go: config translation logic
- Single responsibility: each file has a clear purpose
- No functional changes, same logic and behavior

This improves code organization without changing behavior.
