+++ common/service/frontend/sessions/utils/tools.go
@@ -8,12 +8,10 @@ import (
 	"strings"
 
 	"github.com/gorilla/securecookie"
-	"github.com/gorilla/sessions"
 	"go.uber.org/zap"
 
 	"github.com/pydio/cells/v5/common"
 	"github.com/pydio/cells/v5/common/config"
-	"github.com/pydio/cells/v5/common/config/routing"
 	"github.com/pydio/cells/v5/common/telemetry/log"
 )
 
@@ -67,29 +65,4 @@ func RequestURL(r *http.Request) *url.URL {
 		}
 	}
 	return &u
-}
-
-// SessionOptionsFromConfig creates a default sessions.Options struct using context configuration.
-// It reads the SESSION_TIMEOUT, SameSite policy from config and the REST API path from routing context.
-func SessionOptionsFromConfig(ctx context.Context) *sessions.Options {
-	restApi := routing.RouteIngressURIContext(ctx, common.RouteApiREST, common.DefaultRouteREST)
-	timeout := config.Get(ctx, config.FrontendPluginPath(config.KeyFrontPluginGuiAjax, "SESSION_TIMEOUT")...).Default(60).Int()
-	sameSiteStr := config.Get(ctx, "frontend", "secureCookies", "SameSite").Default("Strict").String()
-
-	var sameSite http.SameSite
-	switch sameSiteStr {
-	case "Lax":
-		sameSite = http.SameSiteLaxMode
-	case "None":
-		sameSite = http.SameSiteNoneMode
-	default:
-		sameSite = http.SameSiteStrictMode
-	}
-
-	return &sessions.Options{
-		Path:     restApi + "/frontend",
-		MaxAge:   60 * timeout,
-		HttpOnly: true,
-		SameSite: sameSite,
-	}
 }
